WIP merge ActiveRelation into ActiveQuery
allow extending only one class to add scopes, fixes #2146

TODO:

- [ ] adjust guide docs
- [ ] adjust README files of extensions
- [ ] finish work and fix test breaks
