+++ extensions/elasticsearch/ActiveRecord.php
@@ -138,19 +138,32 @@ class ActiveRecord extends BaseActiveRecord
 	// TODO add percolate functionality http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-percolate.html
 
 	/**
-	 * @inheritdoc
+	 * Creates an [[ActiveQuery]] instance.
+	 *
+	 * This method is called by [[find()]], [[findBySql()]] to start a SELECT query but also
+	 * by [[hasOne()]] and [[hasMany()]] to create a relational query.
+	 * You may override this method to return a customized query (e.g. `CustomerQuery` specified
+	 * written for querying `Customer` purpose.)
+	 *
+	 * You may also define default conditions that should apply to all queries unless overridden:
+	 *
+	 * ```php
+	 * public static function createQuery()
+	 * {
+	 *     return parent::createQuery()->where(['deleted' => false]);
+	 * }
+	 * ```
+	 *
+	 * Note that all queries should use [[Query::andWhere()]] and [[Query::orWhere()]] to keep the
+	 * default condition. Using [[Query::where()]] will override the default condition.
+	 *
+	 * @param array $config the configuration passed to the ActiveRelation class.
+	 * @return ActiveQuery the newly created [[ActiveQuery]] instance.
 	 */
-	public static function createQuery()
+	public static function createQuery($config)
 	{
-		return new ActiveQuery(['modelClass' => get_called_class()]);
-	}
-
-	/**
-	 * @inheritdoc
-	 */
-	public static function createRelation($config = [])
-	{
-		return new ActiveRelation($config);
+		$config['modelClass'] = get_called_class();
+		return new ActiveQuery($config);
 	}
 
 	// TODO implement copy and move as pk change is not possible
