+++ extensions/mongodb/ActiveQuery.php
@@ -9,9 +9,19 @@ namespace yii\mongodb;
 
 use yii\db\ActiveQueryInterface;
 use yii\db\ActiveQueryTrait;
+use yii\db\ActiveRelationInterface;
+use yii\db\ActiveRelationTrait;
 
 /**
  * ActiveQuery represents a Mongo query associated with an Active Record class.
+ *
+ * An ActiveQuery can be a normal query or be used in a relational context.
+ *
+ * ActiveQuery instances are usually created by [[ActiveRecord::find()]].
+ * Relational queries are created by [[ActiveRecord::hasOne()]] and [[ActiveRecord::hasMany()]].
+ *
+ * Normal Query
+ * ------------
  *
  * ActiveQuery instances are usually created by [[ActiveRecord::find()]].
  *
@@ -25,18 +35,66 @@ use yii\db\ActiveQueryTrait;
  *
  * These options can be configured using methods of the same name. For example:
  *
- * ~~~
+ * ```php
  * $customers = Customer::find()->with('orders')->asArray()->all();
- * ~~~
+ * ```
+ *
+ * Relational query
+ * ----------------
+ *
+ * In relational context ActiveQuery represents a relation between two Active Record classes.
+ *
+ * Relational ActiveQuery instances are usually created by calling [[ActiveRecord::hasOne()]] and
+ * [[ActiveRecord::hasMany()]]. An Active Record class declares a relation by defining
+ * a getter method which calls one of the above methods and returns the created ActiveQuery object.
+ *
+ * A relation is specified by [[link]] which represents the association between columns
+ * of different tables; and the multiplicity of the relation is indicated by [[multiple]].
+ *
+ * If a relation involves a pivot table, it may be specified by [[via()]].
+ * This methods may only be called in a relational context. Same is true for [[inverseOf()]], which
+ * marks a relation as inverse of another relation.
  *
  * @property Collection $collection Collection instance. This property is read-only.
  *
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.0
  */
-class ActiveQuery extends Query implements ActiveQueryInterface
+class ActiveQuery extends Query implements ActiveQueryInterface, ActiveRelationInterface
 {
 	use ActiveQueryTrait;
+	use ActiveRelationTrait;
+
+	/**
+	 * @inheritdoc
+	 */
+	protected function buildCursor($db = null)
+	{
+		if ($this->primaryModel !== null) {
+			// lazy loading
+			if ($this->via instanceof self) {
+				// via pivot collection
+				$viaModels = $this->via->findPivotRows([$this->primaryModel]);
+				$this->filterByModels($viaModels);
+			} elseif (is_array($this->via)) {
+				// via relation
+				/** @var ActiveQuery $viaQuery */
+				list($viaName, $viaQuery) = $this->via;
+				if ($viaQuery->multiple) {
+					$viaModels = $viaQuery->all();
+					$this->primaryModel->populateRelation($viaName, $viaModels);
+				} else {
+					$model = $viaQuery->one();
+					$this->primaryModel->populateRelation($viaName, $model);
+					$viaModels = $model === null ? [] : [$model];
+				}
+				$this->filterByModels($viaModels);
+			} else {
+				$this->filterByModels([$this->primaryModel]);
+			}
+		}
+		return parent::buildCursor($db);
+	}
 
 	/**
 	 * Executes query and returns all results as an array.
