+++ extensions/mongodb/file/ActiveRecord.php
@@ -46,14 +46,31 @@ abstract class ActiveRecord extends \yii\mongodb\ActiveRecord
 {
 	/**
 	 * Creates an [[ActiveQuery]] instance.
-	 * This method is called by [[find()]] to start a "find" command.
-	 * You may override this method to return a customized query (e.g. `ImageFileQuery` specified
-	 * written for querying `ImageFile` purpose.)
+	 *
+	 * This method is called by [[find()]], [[findBySql()]] to start a SELECT query but also
+	 * by [[hasOne()]] and [[hasMany()]] to create a relational query.
+	 * You may override this method to return a customized query (e.g. `CustomerQuery` specified
+	 * written for querying `Customer` purpose.)
+	 *
+	 * You may also define default conditions that should apply to all queries unless overridden:
+	 *
+	 * ```php
+	 * public static function createQuery()
+	 * {
+	 *     return parent::createQuery()->where(['deleted' => false]);
+	 * }
+	 * ```
+	 *
+	 * Note that all queries should use [[Query::andWhere()]] and [[Query::orWhere()]] to keep the
+	 * default condition. Using [[Query::where()]] will override the default condition.
+	 *
+	 * @param array $config the configuration passed to the ActiveRelation class.
 	 * @return ActiveQuery the newly created [[ActiveQuery]] instance.
 	 */
-	public static function createQuery()
+	public static function createQuery($config)
 	{
-		return new ActiveQuery(['modelClass' => get_called_class()]);
+		$config['modelClass'] = get_called_class();
+		return new ActiveQuery($config);
 	}
 
 	/**
@@ -63,18 +80,6 @@ abstract class ActiveRecord extends \yii\mongodb\ActiveRecord
 	public static function getCollection()
 	{
 		return static::getDb()->getFileCollection(static::collectionName());
-	}
-
-	/**
-	 * Creates an [[ActiveRelation]] instance.
-	 * This method is called by [[hasOne()]] and [[hasMany()]] to create a relation instance.
-	 * You may override this method to return a customized relation.
-	 * @param array $config the configuration passed to the ActiveRelation class.
-	 * @return ActiveRelation the newly created [[ActiveRelation]] instance.
-	 */
-	public static function createRelation($config = [])
-	{
-		return new ActiveRelation($config);
 	}
 
 	/**
