+++ extensions/redis/CHANGELOG.md
@@ -7,6 +7,9 @@ Yii Framework 2 redis extension Change Log
 - Bug #1993: afterFind event in AR is now called after relations have been populated (cebe, creocoder)
 - Enh #1773: keyPrefix property of Session and Cache is not restricted to alnum characters anymore (cebe)
 - Chg #2281: Renamed `ActiveRecord::create()` to `populateRecord()` and changed signature. This method will not call instantiate() anymore (cebe)
+- Chg #2146: Removed `ActiveRelation` class and moved the functionality to `ActiveQuery`.
+             All relational queries are now directly served by `ActiveQuery` allowing to use
+             custom scopes in relations (cebe)
 
 2.0.0 alpha, December 1, 2013
 -----------------------------
