+++ extensions/redis/README.md
@@ -148,7 +148,7 @@ class Customer extends \yii\redis\ActiveRecord
 	}
 
 	/**
-	 * @return ActiveRelation defines a relation to the Order record (can be in other database, e.g. elasticsearch or sql)
+	 * @return ActiveQuery defines a relation to the Order record (can be in other database, e.g. elasticsearch or sql)
 	 */
 	public function getOrders()
 	{
