+++ extensions/sphinx/CHANGELOG.md
@@ -5,9 +5,12 @@ Yii Framework 2 sphinx extension Change Log
 ----------------------------
 
 - Bug #1993: afterFind event in AR is now called after relations have been populated (cebe, creocoder)
-- Bug #2160: SphinxQL does not support OFFSET (qiangxue, romeo7)
+- Bug #2160: SphinxQL does not support `OFFSET` (qiangxue, romeo7)
 - Enh #1398: Refactor ActiveRecord to use BaseActiveRecord class of the framework (klimov-paul)
 - Chg #2281: Renamed `ActiveRecord::create()` to `populateRecord()` and changed signature. This method will not call instantiate() anymore (cebe)
+- Chg #2146: Removed `ActiveRelation` class and moved the functionality to `ActiveQuery`.
+             All relational queries are now directly served by `ActiveQuery` allowing to use
+             custom scopes in relations (cebe)
 
 2.0.0 alpha, December 1, 2013
 -----------------------------
