+++ framework/db/ActiveRelationInterface.php
@@ -6,9 +6,10 @@
  */
 
 namespace yii\db;
+use yii\base\InvalidParamException;
 
 /**
- * ActiveRelationInterface defines the common interface to be implemented by active record relation classes.
+ * ActiveRelationInterface defines the common interface to be implemented by relational active record query classes.
  *
  * A class implementing this interface should also use [[ActiveRelationTrait]].
  *
@@ -26,4 +27,14 @@ interface ActiveRelationInterface extends ActiveQueryInterface
 	 * @return static the relation object itself.
 	 */
 	public function via($relationName, $callable = null);
+
+	/**
+	 * Finds the related records for the specified primary record.
+	 * This method is invoked when a relation of an ActiveRecord is being accessed in a lazy fashion.
+	 * @param string $name the relation name
+	 * @param ActiveRecordInterface $model the primary model
+	 * @return mixed the related record(s)
+	 * @throws InvalidParamException if the relation is invalid
+	 */
+	public function findFor($name, $model);
 }
