+++ tests/unit/data/ar/elasticsearch/Customer.php
@@ -63,8 +63,9 @@ class Customer extends ActiveRecord
 
 	}
 
-	public static function createQuery()
+	public static function createQuery($config)
 	{
-		return new CustomerQuery(['modelClass' => get_called_class()]);
+		$config['modelClass'] = get_called_class();
+		return new CustomerQuery($config);
 	}
 }
