+++ tests/unit/data/ar/sphinx/ArticleIndex.php
@@ -25,8 +25,9 @@ class ArticleIndex extends ActiveRecord
 		return $this->source->content;
 	}
 
-	public static function createQuery()
+	public static function createQuery($config)
 	{
-		return new ArticleIndexQuery(['modelClass' => get_called_class()]);
+		$config['modelClass'] = get_called_class();
+		return new ArticleIndexQuery($config);
 	}
 }
